# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PoW CAPTCHA API
  version: 0.5.0
  description: |
    基于 Argon2id + SHA-256 双阶段工作量证明的验证码服务。
    - /challenge：发放挑战
    - /verify：校验解答，返回 captcha_token
    - /siteverify：业务后端核验 token
servers:
  - url: http://localhost:8787
    description: 本地开发
paths:
...
  /api/v1/siteverify:
    post:
      summary: 业务后端核验 captcha_token
      operationId: siteVerify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, secret_key]
              properties:
                token:
                  type: string
                  description: 从 /verify 获取的 captcha_token
                secret_key:
                  type: string
                  description: 站点私钥
      responses:
        '200':
          description: 始终 200，用 success 字段表示结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteVerifyResponse'
...
components:
  schemas:
...
    SiteVerifyResponse:
      type: object
      properties:
        success:
          type: boolean
        challenge_id:
          type: string
        site_key:
          type: string
        error:
          type: string
...